launch script refactoring
